creato file learning.py contenente la parte di apprendimento
